+++ phpmyfaq/admin/assets/src/content/editor.js
@@ -13,236 +13,222 @@
  * @since     2022-01-31
  */
 
-/* Import TinyMCE */
-import tinymce from 'tinymce';
-
-/* Default icons are required. After that, import custom icons if applicable */
-import 'tinymce/icons/default/icons.min.js';
-
-/* Required TinyMCE components */
-import 'tinymce/themes/silver/theme.min.js';
-import 'tinymce/models/dom/model.min.js';
-
-/* Import a skin (can be a custom skin instead of the default) */
-import 'tinymce/skins/ui/oxide/skin.js';
-
-// importing the plugin js.
-import 'tinymce/plugins/advlist';
-import 'tinymce/plugins/anchor';
-import 'tinymce/plugins/autolink';
-import 'tinymce/plugins/link';
-import 'tinymce/plugins/image';
-import 'tinymce/plugins/lists';
-import 'tinymce/plugins/charmap';
-import 'tinymce/plugins/searchreplace';
-import 'tinymce/plugins/wordcount';
-import 'tinymce/plugins/code';
-import 'tinymce/plugins/insertdatetime';
-import 'tinymce/plugins/media';
-import 'tinymce/plugins/nonbreaking';
-import 'tinymce/plugins/table';
-import 'tinymce/plugins/help';
-import 'tinymce/plugins/preview';
-import 'tinymce/plugins/visualblocks';
-import 'tinymce/plugins/pagebreak';
-import 'tinymce/plugins/visualchars';
-import 'tinymce/plugins/codesample';
-import 'tinymce/plugins/save';
-import 'tinymce/plugins/directionality';
-import 'tinymce/plugins/emoticons';
-import 'tinymce/plugins/emoticons/js/emojis';
-import 'tinymce/plugins/fullscreen';
-
-/* content UI CSS is required */
-import contentUiSkinCss from 'tinymce/skins/ui/oxide/content.js';
-
-/* The default content CSS can be changed or replaced with appropriate CSS for the editor content. */
-import contentCss from 'tinymce/skins/content/default/content.js';
+import { Jodit } from 'jodit';
+import 'jodit/esm/plugins/class-span/class-span.js';
+import 'jodit/esm/plugins/clean-html/clean-html.js';
+import 'jodit/esm/plugins/clipboard/clipboard.js';
+import 'jodit/esm/plugins/copy-format/copy-format.js';
+import 'jodit/esm/plugins/delete/delete.js';
+import 'jodit/esm/plugins/fullsize/fullsize.js';
+import 'jodit/esm/plugins/hr/hr.js';
+import 'jodit/esm/plugins/image/image.js';
+import 'jodit/esm/plugins/image-processor/image-processor.js';
+import 'jodit/esm/plugins/image-properties/image-properties.js';
+import 'jodit/esm/plugins/indent/indent.js';
+import 'jodit/esm/plugins/justify/justify.js';
+import 'jodit/esm/plugins/line-height/line-height.js';
+import 'jodit/esm/plugins/media/media.js';
+import 'jodit/esm/plugins/preview/preview.js';
+import 'jodit/esm/plugins/print/print.js';
+import 'jodit/esm/plugins/resizer/resizer.js';
+import 'jodit/esm/plugins/search/search.js';
+import 'jodit/esm/plugins/select/select.js';
+import 'jodit/esm/plugins/source/source.js';
+import 'jodit/esm/plugins/symbols/symbols.js';
+import 'jodit/esm/modules/uploader/uploader.js';
+import 'jodit/esm/plugins/video/video.js';
 
 export const renderEditor = () => {
-  const form = document.getElementById('faqEditor');
-  if (form) {
-    const editorEnabled = form.getAttribute('data-pmf-enable-editor');
-    const editorLanguage = form.getAttribute('data-pmf-editor-language');
-    const defaultLanguage = form.getAttribute('data-pmf-default-url');
-
-    if ('1' !== editorEnabled) {
-      return;
-    }
-
-    tinymce.init({
-      // General options
-      language: editorLanguage,
-      document_base_url: defaultLanguage,
-      skin_url: 'default',
-      content_css: 'default',
-      selector: 'textarea#editor',
-      relative_urls: false,
-      convert_urls: false,
-      remove_linebreaks: false,
-      use_native_selects: true,
-      paste_remove_spans: true,
-      entities: '10',
-      entity_encoding: 'raw',
-      height: '50vh',
-      paste_data_images: true,
-      visualblocks_default_state: true,
-      convert_unsafe_embeds: true,
-      end_container_on_empty_block: true,
-      extended_valid_elements: 'code[class],video[*],audio[*],source[*]',
-      removeformat: [{ selector: '*', attributes: ['style'], split: false, expand: false, deep: true }],
-      importcss_append: true,
-      sandbox_iframes: true,
-
-      // Font handling
-      fontsize_formats: '6pt 8pt 9pt 10pt 11pt 12pt 14pt 16pt 18pt 20pt 24pt 36pt 48pt',
-      font_formats:
-        'Arial=arial,helvetica,sans-serif;' +
-        'Arial Black=arial black,avant garde;' +
-        'Calibri=calibri;' +
-        'Comic Sans MS=comic sans ms,sans-serif;' +
-        'Courier New=courier new,courier;' +
-        'Georgia=georgia,palatino;' +
-        'Helvetica=helvetica;' +
-        'Impact=impact,chicago;' +
-        'Symbol=symbol;' +
-        'Tahoma=tahoma,arial,helvetica,sans-serif;' +
-        'Terminal=terminal,monaco;' +
-        'Times New Roman=times new roman,times;' +
-        'Verdana=verdana,geneva;' +
-        'Webdings=webdings;' +
-        'Wingdings=wingdings,zapf dingbats',
-
-      // Plugins
-      plugins:
-        'advlist autolink link image lists charmap preview anchor pagebreak ' +
-        'searchreplace wordcount visualblocks visualchars code insertdatetime media nonbreaking ' +
-        'save table directionality help emoticons fullscreen',
-
-      emoticons_database: 'emojis',
-
-      // Toolbar
-      menubar: false,
-      toolbar1:
-        'newdocument | undo redo | bold italic underline subscript superscript ' +
-        'strikethrough | styleselect | blocks | fontfamily | fontsize |' +
-        'outdent indent | alignleft aligncenter alignright alignjustify | removeformat |' +
-        'insertfile | cut copy codesample | bullist numlist |' +
-        'link unlink anchor image media | charmap | insertdatetime | table |' +
-        'forecolor backcolor emoticons | searchreplace | ' +
-        'pagebreak | code | fullscreen | preview',
-
-      // Formatting
-      style_formats: [
-        {
-          title: 'Headers',
-          items: [
-            { title: 'h1', block: 'h1' },
-            { title: 'h2', block: 'h2' },
-            { title: 'h3', block: 'h3' },
-            { title: 'h4', block: 'h4' },
-            { title: 'h5', block: 'h5' },
-            { title: 'h6', block: 'h6' },
-          ],
-        },
-
-        {
-          title: 'Blocks',
-          items: [
-            { title: 'p', block: 'p' },
-            { title: 'div', block: 'div' },
-            { title: 'pre', block: 'pre' },
-            { title: 'code', block: 'code' },
-          ],
-        },
-
-        {
-          title: 'Containers',
-          items: [
-            { title: 'blockquote', block: 'blockquote', wrapper: true },
-            { title: 'figure', block: 'figure', wrapper: true },
-          ],
-        },
-      ],
-
-      // File browser
-      file_picker_types: 'image media',
-      file_picker_callback: (callback, value, meta) => {
-        const type = meta.filetype;
-        const w = window,
-          d = document,
-          e = d.documentElement,
-          g = d.getElementsByTagName('body')[0],
-          x = w.innerWidth || e.clientWidth || g.clientWidth,
-          y = w.innerHeight || e.clientHeight || g.clientHeight;
-
-        let mediaBrowser = './media-browser';
-        mediaBrowser += mediaBrowser.indexOf('?') < 0 ? '?type=' + type : '&type=' + type;
-
-        tinymce.activeEditor.windowManager.openUrl({
-          url: mediaBrowser,
-          title: 'Select media file',
-          width: x * 0.8,
-          height: y * 0.8,
-          resizable: 'yes',
-          close_previous: 'no',
-          onMessage: function (api, data) {
-            if (data.mceAction === 'phpMyFAQMediaBrowserAction') {
-              callback(data.url);
-              api.close();
-            }
-          },
-        });
-      },
-
-      // override default upload handler to simulate successful upload
-      images_upload_handler: (blobInfo) =>
-        new Promise((resolve, reject) => {
-          const csrf = document.getElementById('pmf-csrf-token').value;
-          const formData = new FormData();
-          formData.append('file', blobInfo.blob(), blobInfo.filename());
-
-          fetch(`api/content/images?csrf=${csrf}`, {
-            method: 'POST',
-            body: formData,
-            //credentials: 'omit'
-          })
-            .then((response) => {
-              if (response.ok) {
-                return response.json();
-              }
-              throw new Error('Network response was not ok: ', { cause: { response } });
-            })
-            .then((json) => {
-              if (!json || typeof json.location != 'string') {
-                console.log(JSON.stringify(json));
-                throw new Error('Invalid JSON: ' + JSON.stringify(json));
-              }
-              resolve(json.location);
-            })
-            .catch(async (error) => {
-              const errors = await error.cause.response.json();
-              console.log(errors);
-            });
-        }),
-
-      // Custom params
-      csrf: document.getElementById('pmf-csrf-token').value,
-
-      // Image handling
-      image_advtab: true,
-      image_class_list: [
-        { title: 'None', value: '' },
-        { title: 'Responsive', value: 'img-fluid' },
-      ],
-      image_dimensions: true,
-      images_upload_url: '/admin/api/content/images',
-      automatic_uploads: true,
-      setup: (editor) => {
-        editor.on('change', () => {
-          tinymce.triggerSave();
-        });
-      },
-    });
+  const editor = document.getElementById('editor');
+  if (!editor) {
+    return;
   }
+
+  const joditEditor = Jodit.make(editor, {
+    zIndex: 0,
+    readonly: false,
+    activeButtonsInReadOnly: ['source', 'fullsize', 'print', 'about', 'dots'],
+    toolbarButtonSize: 'middle',
+    theme: 'default',
+    saveModeInCookie: false,
+    spellcheck: true,
+    editorCssClass: false,
+    triggerChangeEvent: true,
+    width: 'auto',
+    height: 'auto',
+    minHeight: 100,
+    direction: '',
+    language: 'auto',
+    debugLanguage: false,
+    i18n: 'en',
+    tabIndex: -1,
+    toolbar: true,
+    enter: 'P',
+    defaultMode: Jodit.MODE_WYSIWYG,
+    useSplitMode: false,
+    colors: {
+      greyscale: [
+        '#000000',
+        '#434343',
+        '#666666',
+        '#999999',
+        '#B7B7B7',
+        '#CCCCCC',
+        '#D9D9D9',
+        '#EFEFEF',
+        '#F3F3F3',
+        '#FFFFFF',
+      ],
+      palette: [
+        '#980000',
+        '#FF0000',
+        '#FF9900',
+        '#FFFF00',
+        '#00F0F0',
+        '#00FFFF',
+        '#4A86E8',
+        '#0000FF',
+        '#9900FF',
+        '#FF00FF',
+      ],
+      full: [
+        '#E6B8AF',
+        '#F4CCCC',
+        '#FCE5CD',
+        '#FFF2CC',
+        '#D9EAD3',
+        '#D0E0E3',
+        '#C9DAF8',
+        '#CFE2F3',
+        '#D9D2E9',
+        '#EAD1DC',
+        '#DD7E6B',
+        '#EA9999',
+        '#F9CB9C',
+        '#FFE599',
+        '#B6D7A8',
+        '#A2C4C9',
+        '#A4C2F4',
+        '#9FC5E8',
+        '#B4A7D6',
+        '#D5A6BD',
+        '#CC4125',
+        '#E06666',
+        '#F6B26B',
+        '#FFD966',
+        '#93C47D',
+        '#76A5AF',
+        '#6D9EEB',
+        '#6FA8DC',
+        '#8E7CC3',
+        '#C27BA0',
+        '#A61C00',
+        '#CC0000',
+        '#E69138',
+        '#F1C232',
+        '#6AA84F',
+        '#45818E',
+        '#3C78D8',
+        '#3D85C6',
+        '#674EA7',
+        '#A64D79',
+        '#85200C',
+        '#990000',
+        '#B45F06',
+        '#BF9000',
+        '#38761D',
+        '#134F5C',
+        '#1155CC',
+        '#0B5394',
+        '#351C75',
+        '#733554',
+        '#5B0F00',
+        '#660000',
+        '#783F04',
+        '#7F6000',
+        '#274E13',
+        '#0C343D',
+        '#1C4587',
+        '#073763',
+        '#20124D',
+        '#4C1130',
+      ],
+    },
+    colorPickerDefaultTab: 'background',
+    imageDefaultWidth: 300,
+    removeButtons: [],
+    disablePlugins: [],
+    extraButtons: [],
+    buttons: [
+      'source',
+      '|',
+      'bold',
+      'strikethrough',
+      'underline',
+      'italic',
+      '|',
+      'ul',
+      'ol',
+      'superscript',
+      'subscript',
+      '|',
+      'justify',
+      'outdent',
+      'indent',
+      '|',
+      'font',
+      'fontsize',
+      'lineHeight',
+      'brush',
+      'paragraph',
+      '|',
+      'copy',
+      'cut',
+      'paste',
+      'selectall',
+      '|',
+      'file',
+      'image',
+      'video',
+      'table',
+      'link',
+      '|',
+      'undo',
+      'redo',
+      '|',
+      'classSpan',
+      'hr',
+      'eraser',
+      'copyformat',
+      '|',
+      'symbols',
+      'fullsize',
+      'preview',
+      'print',
+    ],
+    events: {},
+    textIcons: false,
+    uploader: {
+      url: '/admin/api/content/images?csrf=' + document.getElementById('pmf-csrf-token').value,
+      format: 'json',
+      isSuccess: (response) => {
+        return !response.error;
+      },
+      getMessage: (response) => {
+        return response.msg;
+      },
+    },
+    filebrowser: {
+      ajax: {
+        url: '/admin/api/media-browser',
+        contentType: 'application/json; charset=UTF-8',
+      },
+      createNewFolder: false,
+      deleteFolder: false,
+      moveFolder: false,
+      showFoldersPanel: false,
+      showFileSize: true,
+      showFileName: true,
+    },
+  });
 };